requirements.txt
sqlalchemy~=2.0.7
pydantic~=1.10.4
sqlalchemy-utils==0.40.0
passlib
fastapi==0.95.0
pyhumps
pymysql
uvicorn==0.21.1
gunicorn
starlette==0.26.1
twilio==7.17.0
mailjet_rest==1.3.4
email-validator
python-multipart
bcrypt
requests~=2.28.2
PyJWT~=2.6.0
pyotp~=2.8.0